refactor(shell): Surface 端口的文档 handler 抽出巨壳——壳只接线，取数规则归模块
修「文稿为空」时给 NomiStudioApp 加了正文回落逻辑，把这个已在白名单里的巨壳从 903 顶到 915。
基线只减不增（R12），所以不是抬基线而是拆：读/写文档这两条 handler 有自己的取数规则
（编辑器是活视图 vs 持久 contentJson 是真相源；写必须落在真实 ProseMirror 位置上），
是一个内聚单元，搬进 project/documentSurfaceHandlers.ts。画布/时间线/资产那几条留在壳里
——它们只是转调既有执行器，没有自己的取数规则，搬走反而多一层。

903 → 883，棘轮锁到新低点。workbench 3635 条测试全过。
