+++ src/workbench/NomiStudioApp.tsx
@@ -22,8 +22,7 @@ import {
 } from './generationCanvas/agent/proposalUndo'
 import { useGenerationCanvasStore } from './generationCanvas/store/generationCanvasStore'
 import { readGenerationCanvasSnapshot } from './generationCanvas/agent/generationCanvasTools'
-import { docToPlainText } from './generationCanvas/runner/textActions'
-import type { TiptapDocJson } from './generationCanvas/model/generationCanvasTypes'
+import { readDocumentSurface, writeDocumentSurface } from './project/documentSurfaceHandlers'
 import {
   captureCanvasDeleteRawEvidence,
   captureCanvasWriteRawEvidence,
@@ -280,39 +279,8 @@ export default function NomiStudioApp(): JSX.Element {
       registerProjectCanvasReadSurface(
         projectSurface,
         readGenerationCanvasSnapshot,
-        ({ documentId, scope }) => {
-          const store = useWorkbenchStore.getState()
-          if (store.activeDocumentId !== documentId) {
-            throw new SurfacePortWireError('surface_port_stale')
-          }
-          const tools = store.creationDocumentTools
-          if (tools) return { text: scope === 'full' ? tools.readFullText() : tools.readSelectionText() }
-          // 编辑器是正文的**活视图**，不是正文的唯一真相源——正文持久化在 workbenchDocuments[].contentJson。
-          // 分镜页/生成页不挂载 WorkbenchEditor，此前这里一律判 stale，Agent 于是把「读不到」讲成
-          // 「当前创作区文稿为空」（2026-09-03 真实付费闭环走查实测）。能力的可用性不该绑在某个 UI
-          // 组件的挂载生命周期上。选区是编辑器**独有**的状态，没有编辑器就没有选区——那种情况仍判
-          // stale，不拿全文冒充选区。
-          if (scope === 'full') {
-            const document = store.workbenchDocuments.find((item) => item.id === documentId)
-            if (document) return { text: docToPlainText(document.contentJson as TiptapDocJson | undefined) }
-          }
-          throw new SurfacePortWireError('surface_port_stale')
-        },
-        ({ documentId, operation, content, target, preconditions, signal, assertCurrent }) => {
-          const store = useWorkbenchStore.getState()
-          const tools = store.creationDocumentTools
-          if (!tools || store.activeDocumentId !== documentId) {
-            throw new SurfacePortWireError('surface_port_stale')
-          }
-          if (signal.aborted) throw new SurfacePortWireError('capability_cancelled')
-          assertCurrent()
-          return tools.applyDocumentWrite({
-            operation,
-            content,
-            target: target as never,
-            preconditions: preconditions as never,
-          })
-        },
+        readDocumentSurface,
+        writeDocumentSurface,
         ({ operation, input, nodeId }) => {
           try {
             if (operation === 'delete_canvas_nodes') {
